Adding function definitions and calls
